Separate default template from nix template
Previously `templates.default` was just a synonym for `templates.nix`, but
recently these two have diverged. The nix template now leans into things like
Project Manager and eschews giving any alternatives for building. This is
primarily meant to be used for projects that are inseparable from Nix (e.g.,
Project Manager itself, this repo, dotfiles, etc.). The default template is just
the common basis of every template, and expects to be used for projects that
allow a non-Nix build (but for which we don’t know what that non-Nix build is.