Have `sync-template` reference the local template
Previously, `sync-template` always looked up the template at
`github:sellout/flaky`. This was annoying if you were trying to test changes on
a branch or something – changing the flake in `nix run <flake>#sync-template`
would have no effect on the template used.

Now building the template replaces the upstream location with the flake store
path, using the template that corresponds to the location of `sync-template`
itself.

This also removes a workaround from when `wrapProgram` didn’t play nice with
strict Bash.