Remove templated versions
The template fields should represent things that don’t change over time (so we
can re-run it when we update the templates).

Versions are not like this – they always start with the same value (so there’s
no benefit to a template in the first place), and they are expected to change.