Adopt nix-systems
This adopts the approach described in
https://github.com/nix-systems/nix-systems#readme for specifying flake systems.

All templates have been updated to use it, and this also adds
garnix-io/nix-systems as an input rather than hard-coding a list of the systems
that garnix supports.

`defaultSystems` is still defined in the `lib` until the packages that form a
cycle with this one (bash-strict-mode and Project Manager) are updated to use
nix-systems as well.