Remove templates & dev shells
They now live in sellout/flaky-environments. The reasons as listed in that repo:

1. It is the library that most projects rely on. The components here are for
bootstrapping new projects or adding minimal Nix support to existing ones.
2. It reduces the surface area of Flaky’s cyclic-dependency issues[^1].
3. Dependencies of templates induce new cycles, but they’re superfluous, so this
eliminates those.

An additional benefit is that it removes the `__noChroot` checks from Flaky,
allowing the automatic flake updates to be auto-merged (assuming all the checks
pass).

This change also contains fixes to `validate-emplate` to support templates in
other repos, and to fix the bug that the Haskell template was running into.

[^1]: Flaky is the basis of _every_ project of mine. Any projects of mine that
Flaky depends on (currently Bash strict mode and Project Manager) results in a
cyclic dependency between Flaky and that project. Nix is actually quite good at
managing cyclic dependencies (yay, laziness), but it does introduce various
complexities, so minimizing what’s involved with the cycles is helpful.